examples/requirements.txt
solmate-sdk
paho-mqtt